+++ tests/unit/algo/action_classification/test_movinet.py
@@ -9,7 +9,7 @@ from otx.algo.utils.support_otx_v1 import OTXv1Helper
 class TestMoViNet:
     @pytest.fixture()
     def fxt_movinet(self) -> MoViNet:
-        return MoViNet(num_classes=10)
+        return MoViNet(label_info=10)
 
     def test_load_from_otx_v1_ckpt(self, fxt_movinet, mocker) -> None:
         mock_load_ckpt = mocker.patch.object(OTXv1Helper, "load_action_ckpt")
