+++ tests/unit/algo/action_classification/test_x3d.py
@@ -9,8 +9,8 @@ from otx.algo.utils.support_otx_v1 import OTXv1Helper
 
 class TestX3D:
     @pytest.fixture()
-    def fxt_x3d(self) -> X3D:
-        return X3D(num_classes=10)
+    def fxt_x3d(self, fxt_multiclass_labelinfo) -> X3D:
+        return X3D(label_info=10)
 
     def test_load_from_otx_v1_ckpt(self, fxt_x3d, mocker):
         mock_load_ckpt = mocker.patch.object(OTXv1Helper, "load_action_ckpt")
