+++ tests/unit/algo/action_detection/test_x3d_fastrcnn.py
@@ -10,7 +10,7 @@ from otx.algo.utils.support_otx_v1 import OTXv1Helper
 class TestX3DFastRCNN:
     @pytest.fixture()
     def fxt_x3d_fast_rcnn(self) -> X3DFastRCNN:
-        return X3DFastRCNN(num_classes=10, topk=3)
+        return X3DFastRCNN(label_info=10, topk=3)
 
     def test_load_from_otx_v1_ckpt(self, fxt_x3d_fast_rcnn, mocker):
         mock_load_ckpt = mocker.patch.object(OTXv1Helper, "load_action_ckpt")
