+++ tests/unit/algo/classification/test_torchvision_model.py
@@ -8,7 +8,7 @@ from otx.core.types.export import TaskLevelExportParameters
 
 @pytest.fixture()
 def fxt_tv_model():
-    return OTXTVModel(backbone="mobilenet_v3_small", num_classes=10)
+    return OTXTVModel(backbone="mobilenet_v3_small", label_info=10)
 
 
 class TestOTXTVModel:
@@ -47,6 +47,6 @@ class TestOTXTVModel:
         assert outputs.has_xai_outputs == explain_mode
 
     def test_freeze_backbone(self):
-        freezed_model = OTXTVModel(backbone="resnet50", num_classes=10, freeze_backbone=True)
+        freezed_model = OTXTVModel(backbone="resnet50", label_info=10, freeze_backbone=True)
         for param in freezed_model.model.backbone.parameters():
             assert not param.requires_grad
