+++ tests/unit/algo/instance_segmentation/heads/test_custom_rtmdet_ins_head.py
@@ -68,7 +68,7 @@ class TestCustomRTMDetInsSepBNHead:
         )
 
     def test_predict_by_feat_ov(self, tmpdir) -> None:
-        lit_module = RTMDetInst(num_classes=1, variant="tiny")
+        lit_module = RTMDetInst(label_info=1, variant="tiny")
         exported_model_path = lit_module.export(
             output_dir=Path(tmpdir),
             base_name="exported_model",
