+++ tests/unit/core/model/test_base.py
@@ -22,12 +22,12 @@ class MockNNModule(torch.nn.Module):
 class TestOTXModel:
     def test_smart_weight_loading(self, mocker) -> None:
         with mocker.patch.object(OTXModel, "_create_model", return_value=MockNNModule(2)):
-            prev_model = OTXModel(num_classes=2)
+            prev_model = OTXModel(label_info=2)
             prev_model.label_info = ["car", "truck"]
             prev_state_dict = prev_model.state_dict()
 
         with mocker.patch.object(OTXModel, "_create_model", return_value=MockNNModule(3)):
-            current_model = OTXModel(num_classes=3)
+            current_model = OTXModel(label_info=3)
             current_model.classification_layers = ["model.head.weight", "model.head.bias"]
             current_model.classification_layers = {
                 "model.head.weight": {"stride": 1, "num_extra_classes": 0},
@@ -57,7 +57,7 @@ class TestOTXModel:
         mock_main_scheduler = mocker.create_autospec(spec=torch.optim.lr_scheduler.LRScheduler)
 
         with mocker.patch.object(OTXModel, "_create_model", return_value=MockNNModule(3)):
-            current_model = OTXModel(num_classes=3)
+            current_model = OTXModel(label_info=3)
 
         mock_trainer = mocker.create_autospec(spec=Trainer)
         mock_trainer.lr_scheduler_configs = [
